requirements.txt
# 3DN Scraper - Python Dependencies
# Python 3.11+ required

# Core web scraping
crawl4ai==0.7.6
firecrawl-py==1.12.0

# HTTP requests
aiohttp>=3.9.0
requests>=2.31.0

# Data handling
python-dotenv>=1.0.0

# ==================== API Dependencies ====================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.1  # Pin to compatible version with passlib
python-multipart>=0.0.6

# Job Queue & Background Tasks
celery>=5.3.0
redis>=5.0.0

# Scheduling
apscheduler>=3.10.0

# HTTP Client (for webhooks)
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=20.0.0
